refactor(core): create subdir for vocabs
Will then split up words into smaller "groups".
